requirements.txt
streamlit
pandas
plotly
openai
openpyxl
fpdf2
kaleido
